Add distinction between failed and waiting MultiEdge-indices.
Before this, the index-getting functions only returned `Nothing`s
which where interpreted as `Wait` by the supremum computation on the
monadic graph. Thus cases where the index actually fails were not propagated
as such and the computation could not fail properly.

This fixes #28.

(Also change order of monad transformers. Now the log is also kept in
case of an error.)