Reintroduce finer loop type computation (#201).
The problem is the distinction between const / non-const types.
We had a similar computation before, but it was disabled in #105,
probably because too many constraints were left unsolved and it was too
much hassle to deal with it at the time.

We mostly reinstantiate the previous behaviour, it is now:
```
body_in := nonConst(body_out)
loop_in ~<=~ body_in
loop_out := body_in
```
The difference now is that the "subtyping" of `loop_in` and `body_in`
is not *real* subtyping. Instead we have a new constraint
`UnifyWithConstSubtype`, which is like unification on all constructors
except `Const`/`NonConst`, where it is like subtyping instead. This makes it
stricter than normal subtyping, which we used before.