Throw unification error (not `impossible`) for non-idempotent subs.
This should fix the currently last problem with #216.

What happened is that supremum computation encountered statements like
`sup{a,a+1} = b` where `a b : Constness`. Solving this using the monadic-graph
machinery, resulted in constraints of the form `a = a + 1` which were created
in a temporary constraint set, and the fact that this constraint has no solution
should be further used in the graph-search. But since the error was
`impossible`, it was not catched as a "possible, valid exception" and forwarded
to the user.